move all state to higher component app.js
