+++ src/Components/FeedbackContent.jsx
@@ -1,10 +1,7 @@
 import React, { useState, useEffect } from "react";
 import LayoutStyles from "./styles/layout.module.css";
 import Graphic from "./Profile/Graphic";
-import axios from "axios";
 import Dropdown from "./Dropdown";
-
-import { getFeedbacks } from "./Helpers/getters";
 
 {
   /* <Graphic type="scatter" xLabel="x" yLabel="y" title="Title" />
@@ -15,63 +12,22 @@ import { getFeedbacks } from "./Helpers/getters";
 <Graphic type="swarm" title="Title" /> */
 }
 
-function FeedbackContent() {
-  const [surveys, setSurveys] = useState([]);
-  const [virtues, setVirtues] = useState([]);
-  const [virtueBuckets, setVirtueBuckets] = useState([]);
-  const [employees, setEmployees] = useState([]);
-  const [graphSettings, setGraphSettings] = useState({
-    timeline: [2020, "Courage"],
-    pie: [2018, "Stephen Khan"],
-    quadrant: [2019],
-    scatter: [2019],
-    candle: [2019, "Execution", "Stephen Khan"],
-    bar: [2020],
-  });
-
-  useEffect(() => {
-    Promise.all([
-      axios.get("/surveys"),
-      axios.get("/virtues"),
-      axios.get("/virtues/buckets"),
-      axios.get("/employees"),
-    ]).then((all) => {
-      setSurveys(all[0].data);
-      setVirtues(all[1].data);
-      setVirtueBuckets(all[2].data);
-      setEmployees(all[3].data);
-    });
-  }, []);
-
-  let feedbacks = [];
-  if (
-    surveys.length > 0 &&
-    virtues.length > 0 &&
-    virtueBuckets.length > 0 &&
-    employees.length > 0
-  ) {
-    feedbacks = getFeedbacks(surveys, virtues, virtueBuckets, employees);
-  }
-
+function FeedbackContent(props) {
+  console.log(props);
+  const { feedbacks, virtueBuckets, graphSettings, setGraphSettings } = props;
   const yearOptions = [2017, 2018, 2019, 2020];
-  const virtueBucketOptions = [...virtueBuckets].map((data) => data.name);
+  const virtueBucketOptions = [...props.virtueBuckets].map((data) => data.name);
 
-  const handleYearChange = (chart, event) => {
+  const handleChange = (chart, event, bucket = false) => {
     const temp = [...graphSettings[chart]];
 
-    temp.splice(0, 1, event.target.value);
+    bucket
+      ? temp.splice(1, 1, event.target.getAttribute("value"))
+      : temp.splice(0, 1, event.target.value);
 
     setGraphSettings({ ...graphSettings, [chart]: temp });
   };
 
-  const handleBucketChange = (event, chart) => {
-    const temp = [...graphSettings[chart]];
-
-    temp.splice(1, 1, event.target.getAttribute("value"));
-
-    setGraphSettings({ ...graphSettings, [chart]: temp });
-  };
-  
   return (
     <div className={LayoutStyles.main_heading}>
       <h3>DASHBOARD</h3>
@@ -84,7 +40,7 @@ function FeedbackContent() {
           <Dropdown
             title="Select Year"
             options={yearOptions}
-            onClick={(event) => handleYearChange("bar", event)}
+            onClick={(event) => handleChange("bar", event)}
           />
           {feedbacks.length > 0 && (
             <Graphic
@@ -105,7 +61,7 @@ function FeedbackContent() {
           <Dropdown
             title="Select Year"
             options={yearOptions}
-            onClick={(event) => handleYearChange("quadrant", event)}
+            onClick={(event) => handleChange("quadrant", event)}
           />
           {feedbacks.length > 0 && (
             // <Graphic
@@ -117,13 +73,10 @@ function FeedbackContent() {
             //   settings={graphSettings}
             // />
             <Graphic
-
               type="quadrant"
               xLabel="x"
               yLabel="y"
-
               type="radar"
-
               title="Title"
               feedbacks={feedbacks}
               settings={graphSettings}
@@ -138,12 +91,12 @@ function FeedbackContent() {
           <Dropdown
             title="Select Year"
             options={yearOptions}
-            onClick={(event) => handleYearChange("candle", event)}
+            onClick={(event) => handleChange("candle", event)}
           />
           <Dropdown
             title="Select Virtue Bucket"
             options={virtueBucketOptions}
-            onClick={(event) => handleBucketChange(event, "candle")}
+            onClick={(event) => handleChange("candle", event, true)}
           />
           {feedbacks.length > 0 && (
             <Graphic
@@ -163,12 +116,12 @@ function FeedbackContent() {
           <Dropdown
             title="Select Year"
             options={yearOptions}
-            onClick={(event) => handleYearChange("timeline", event)}
+            onClick={(event) => handleChange("timeline", event)}
           />
           <Dropdown
             title="Select Virtue Bucket"
             options={virtueBucketOptions}
-            onClick={(event) => handleBucketChange(event, "timeline")}
+            onClick={(event) => handleChange("timeline", event, true)}
           />
           {feedbacks.length > 0 && (
             <Graphic
