--- src/Components/Home.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-import LayoutStyles from "./styles/layout.module.css";
-
-function Home() {
-  return (
-    <div>
-      <div>Home</div>
-    </div>
-  );
-}
-
-export default Home;
